Move Generation for pawns
Implemented Move Generation for White and Black Pawns. This includes:
- Regular pawn forward move
- Pawn start (double forward)
- Pawn capture left and right
- Pawn promotion forward
- Pawn capture promotion left and right
- Pawn En Passant capture left and right
At the end of the Move Generator, we have a List containing all the possible moves for the given side in the current position.
Implemented a Print MoveList function that prints all the possible moves to the console, including all the relevant data for each move.