Finished Iterative Deepening function
The engine now uses a time limit of 10 seconds to search the position for the best move, instead of searching to a set depth.
This time limit is customizable, and it is designed to improve time consistency between different moves, as well as to improve search quality when ordering is very high and we can search to higher depths.