Implemented transposition table
The transposition table is used to speed up the search algorithm. It stores previously explored moves and positions in a hash table which is then probed during search in order to find previous searches on the same position, and to reuse that exploration. It greatly increases the speed at which the engine can find moves, and allows the game to search to higher depths.